+++ dart/lib/src/platform_checker.dart
@@ -3,7 +3,7 @@ import 'platform/platform.dart';
 /// Helper to check in which enviroment the library is running.
 /// The envirment checks (release/debug/profile) are mutually exclusive.
 class PlatformChecker {
-  PlatformChecker({
+  const PlatformChecker({
     this.platform = instance,
     this.isWeb = identical(0, 0.0),
   });