+++ flutter/test/mocks.dart
@@ -1,6 +1,7 @@
 import 'package:mockito/annotations.dart';
 import 'package:sentry/sentry.dart';
 import 'package:sentry/src/platform/platform.dart';
+import 'package:sentry/src/platform_checker.dart';
 
 const fakeDsn = 'https://abc@def.ingest.sentry.io/1234567';
 
@@ -62,4 +63,39 @@ class MockPlatform implements Platform {
 
   @override
   bool get isFuchsia => (operatingSystem == 'fuchsia');
+}
+
+class MockPlatformChecker implements PlatformChecker {
+  MockPlatformChecker({
+    this.isDebug = false,
+    this.isProfile = false,
+    this.isRelease = false,
+    this.isWebValue = false,
+    this.hasNativeIntegration = false,
+    Platform? mockPlatform,
+  }) : _mockPlatform = mockPlatform ?? MockPlatform();
+
+  final bool isDebug;
+  final bool isProfile;
+  final bool isRelease;
+  final bool isWebValue;
+  late final Platform _mockPlatform;
+
+  @override
+  bool hasNativeIntegration = false;
+
+  @override
+  bool isDebugMode() => isDebug;
+
+  @override
+  bool isProfileMode() => isProfile;
+
+  @override
+  bool isReleaseMode() => isRelease;
+
+  @override
+  bool get isWeb => isWebValue;
+
+  @override
+  Platform get platform => _mockPlatform;
 }