+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Unreleased
 
+* Feat: Enrich events with more context (#452)
 * Feat: Add Culture Context (#491)
 * Feat: `beforeSend` callback accepts async code (#494)